requirements.txt
torch
torchvision
torchaudio
soundfile
numpy==1.26.4
torch_audiomentations
tqdm
matplotlib
pandas
wandb
comet_ml
hydra-core

black
isort
pre-commit

datasets
wget
comet-ml
pytorch_lightning